Attempted and then commented out jellyfish compat stuff
It's causing a weird-ass lag problem. I might return to this when jellyfish are more complete and just PR to UA if I can't fix the problem, but in any case this seems out-of-scope for the potions as they only change entities that have naturally existing small and large forms, which, to my knowledge, jellyfish don't.